+++ internal/util/conv.go
@@ -2,9 +2,9 @@ package util
 
 import "encoding/binary"
 
-func Itob(v int64) []byte {
+func Itob(v uint64) []byte {
 	b := make([]byte, 8)
-	binary.BigEndian.PutUint64(b, uint64(v))
+	binary.BigEndian.PutUint64(b, v)
 
 	return b
 }
